audit followups: hot-add tracks, atomic save, perf+memory fixes
Bundle of changes from the 4-agent audit pass + race recording work,
applied incrementally across the prior overnight run:

- replay: pointer_mem int → int32 (~77 MB host RAM saved on 1M buf)
- train: atomic checkpoint write (tmp + os.replace) — prevents 0-byte
  _final.pt on SIGTERM mid-write (seen twice on overnight saves)
- train: act() uses num_tau=1 — argmax over Q discards multi-τ noise
  in expectation, ~16× faster rollout forward
- reward: wall_penalty 0.05 → 0.02 — anti-backward-driving trap (50 Hz
  × 0.05 = 2.5/s sustained wall-grind out-weighed checkpoint rewards
  ~80:1, biased policy toward methodology-flagged local optimum)
- track_sampler: min_samples_per_track=10 cold-start floor — uniform
  among least-sampled tracks until everyone clears the floor; without
  it, 2-episode vs 30-episode tracks get 100× sampling skew from EMA
  variance alone
- track_sampler.update: silent skip on unknown slug — race-safe with
  crash-counter remove_track in another thread; previously KeyError'd
  rollout threads dead
- train: hot-add-tracks via 60s savestate-dir poll + blacklisted_tracks
  set — record new track .sav files into data/savestates/ during a live
  run, they join the curriculum without restart
- dolphin_env: multi-savestate glob ({slug}_*.sav, random.choice)
- record_savestates: 12 saves per track at distributed race_completion
  thresholds [1.0..2.833] — replaces single-savestate-per-track flow
- model: num_tau override param on forward/q_values (enables act()
  num_tau=1 while learn_step uses cfg.num_tau)
- pyproject: requires-python = ">=3.12,<3.14" — Linux Dolphin built
  from source links libpython3.12, prior 3.13-only lock broke Vast

+++ src/mkw_rl/rl/replay.py
@@ -294,18 +294,21 @@ class PER:
         )
 
         # pointer_mem: each entry holds indices into state_mem for the stack
-        # and n_stack, plus indices into reward_mem for the n-step reward window.
+        # and n_stack, plus indices into reward_mem for the n-step reward
+        # window. int32 not int64 — storage_size <= 1.75M < 2^31, so 4 bytes
+        # per pointer suffices and saves ~77 MB host RAM at prod scale
+        # (1.75M slots × 11 ints × 4 bytes vs 8 bytes).
         self.trans_dtype = np.dtype(
             [
-                ("state", int, self.framestack),
-                ("n_state", int, self.framestack),
-                ("reward", int, self.n_step),
+                ("state", np.int32, self.framestack),
+                ("n_state", np.int32, self.framestack),
+                ("reward", np.int32, self.n_step),
             ]
         )
         self.blank_trans = (
-            np.zeros(self.framestack, dtype=int),
-            np.zeros(self.framestack, dtype=int),
-            np.zeros(self.n_step, dtype=int),
+            np.zeros(self.framestack, dtype=np.int32),
+            np.zeros(self.framestack, dtype=np.int32),
+            np.zeros(self.n_step, dtype=np.int32),
         )
         self.pointer_mem = np.array([self.blank_trans] * size, dtype=self.trans_dtype)
 
@@ -359,9 +362,9 @@ class PER:
             reward_array = self.reward_buffer[stream][: self.n_step]
 
             self.pointer_mem[self.point_mem_idx] = (
-                np.array(state_array, dtype=int),
-                np.array(n_state_array, dtype=int),
-                np.array(reward_array, dtype=int),
+                np.array(state_array, dtype=np.int32),
+                np.array(n_state_array, dtype=np.int32),
+                np.array(reward_array, dtype=np.int32),
             )
             # self._set_priority_min(self.point_mem_idx, sqrt(self.max_prio))
             self.st.append(self.max_prio**self.alpha)
@@ -392,9 +395,9 @@ class PER:
                 reward_array.extend([0])
 
             self.pointer_mem[self.point_mem_idx] = (
-                np.array(first_array, dtype=int),
-                np.array(second_array, dtype=int),
-                np.array(reward_array, dtype=int),
+                np.array(first_array, dtype=np.int32),
+                np.array(second_array, dtype=np.int32),
+                np.array(reward_array, dtype=np.int32),
             )
             self.st.append(self.max_prio**self.alpha)
